task-service: architecture refactoring: delete modules and add liquibase in docker-compose
