+++ backend/settings.gradle
@@ -1,7 +1,3 @@
 rootProject.name = 'backend'
 
-include 'test-service'
 include 'task-service'
-include 'task-service:api'
-findProject(':task-service:api')?.name = 'api'
-
