--- backend/task-service/api/src/main/java/ru/calendorny/taskservice/dto/response/ApiErrorReponse.java
@@ -1,18 +1,0 @@
-package ru.calendorny.taskservice.dto.response;
-
-import lombok.Data;
-import java.util.List;
-
-@Data
-public class ApiErrorReponse {
-
-    String description;
-
-    String code;
-
-    String exceptionName;
-
-    String exceptionMessage;
-
-    List<String> stackTrace;
-}
