--- backend/task-service/api/src/main/java/ru/calendorny/taskservice/dto/response/ValidationErrorResponse.java
@@ -1,21 +1,0 @@
-package ru.calendorny.taskservice.dto.response;
-
-import lombok.Data;
-import ru.calendorny.taskservice.util.ValidationError;
-import java.util.List;
-
-@Data
-public class ValidationErrorResponse {
-
-    String description;
-
-    String code;
-
-    String exceptionName;
-
-    String exceptionMessage;
-
-    List<String> stacktrace;
-
-    List<ValidationError> validationErrors;
-}
