--- backend/task-service/api/src/main/java/ru/calendorny/taskservice/enums/RecurrenceFrequency.java
@@ -1,7 +1,0 @@
-package ru.calendorny.taskservice.enums;
-
-public enum RecurrenceFrequency {
-
-    DAILY, WEEKLY, MONTHLY, YEARLY
-
-}
