--- backend/task-service/api/src/main/java/ru/calendorny/taskservice/enums/TaskStatus.java
@@ -1,7 +1,0 @@
-package ru.calendorny.taskservice.enums;
-
-public enum TaskStatus {
-
-    PENDING, COMPLETED, CANCELLED
-
-}
