--- backend/task-service/api/src/main/java/ru/calendorny/taskservice/enums/WeekDay.java
@@ -1,7 +1,0 @@
-package ru.calendorny.taskservice.enums;
-
-public enum WeekDay {
-
-    MO, TU, WE, TH, FR, SA, SU
-
-}
