+++ backend/task-service/build.gradle
@@ -1,5 +1,8 @@
 dependencies {
-
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.3'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.postgresql:postgresql:42.7.5'
+    implementation 'org.liquibase:liquibase-core:4.31.1'
 }
 
 tasks.register("prepareKotlinBuildScriptModel") {
