--- backend/task-service/impl/src/main/java/ru/calendorny/taskservice/TaskServiceApplication.java
@@ -1,13 +1,0 @@
-package ru.calendorny.taskservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TaskServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TaskServiceApplication.class, args);
-    }
-
-}
