--- backend/task-service/impl/src/main/resources/application.yaml
@@ -1,19 +1,0 @@
-spring:
-  application:
-    name: task-service
-  datasource:
-    url: jdbc:postgresql://localhost:5432/task_service_db
-    username: user
-    password: password
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-  liquibase:
-    enabled: true
-    change-log: classpath:/db/changelog/db-changelog-master.yml
