--- backend/task-service/impl/src/test/java/ru/calendorny/taskservice/TaskServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.calendorny.taskservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TaskServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
