+++ backend/task-service/src/main/java/ru/calendorny/taskservice/TaskServiceApplication.java
@@ -1,0 +1,12 @@
+package ru.calendorny.taskservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TaskServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TaskServiceApplication.class, args);
+    }
+}
