+++ backend/task-service/src/main/java/ru/calendorny/taskservice/dto/request/CreateTaskRequest.java
@@ -1,0 +1,22 @@
+package ru.calendorny.taskservice.dto.request;
+
+import java.time.LocalDate;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import ru.calendorny.taskservice.enums.TaskStatus;
+import ru.calendorny.taskservice.util.RecurrenceRule;
+
+@Getter
+@EqualsAndHashCode
+public class CreateTaskRequest {
+
+    private String title;
+
+    private String description;
+
+    private LocalDate due_date;
+
+    private TaskStatus status;
+
+    private RecurrenceRule recurrenceRule;
+}
