+++ backend/task-service/src/main/java/ru/calendorny/taskservice/enums/RecurrenceFrequency.java
@@ -1,0 +1,8 @@
+package ru.calendorny.taskservice.enums;
+
+public enum RecurrenceFrequency {
+    DAILY,
+    WEEKLY,
+    MONTHLY,
+    YEARLY
+}
