+++ backend/task-service/src/main/java/ru/calendorny/taskservice/enums/TaskStatus.java
@@ -1,0 +1,7 @@
+package ru.calendorny.taskservice.enums;
+
+public enum TaskStatus {
+    PENDING,
+    COMPLETED,
+    CANCELLED
+}
