+++ backend/task-service/src/main/java/ru/calendorny/taskservice/enums/WeekDay.java
@@ -1,0 +1,11 @@
+package ru.calendorny.taskservice.enums;
+
+public enum WeekDay {
+    MO,
+    TU,
+    WE,
+    TH,
+    FR,
+    SA,
+    SU
+}
