+++ backend/task-service/src/main/java/ru/calendorny/taskservice/util/RecurrenceRule.java
@@ -1,0 +1,25 @@
+package ru.calendorny.taskservice.util;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import ru.calendorny.taskservice.enums.RecurrenceFrequency;
+import ru.calendorny.taskservice.enums.WeekDay;
+
+public class RecurrenceRule {
+
+    RecurrenceFrequency freq;
+
+    Integer interval;
+
+    LocalDateTime until;
+
+    Integer count;
+
+    List<WeekDay> byDay;
+
+    List<Integer> byMonthDay;
+
+    List<Integer> byMnoth;
+
+    WeekDay weekStart;
+}
