+++ backend/task-service/src/test/java/ru/calendorny/taskservice/TaskServiceApplicationTests.java
@@ -1,0 +1,11 @@
+package ru.calendorny.taskservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TaskServiceApplicationTests {
+
+    @Test
+    void contextLoads() {}
+}
